Issue #48. Refactoring
